% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortCharacterization.R
\name{getCharacterizationResultSet}
\alias{getCharacterizationResultSet}
\title{Get Characterization Result Set}
\usage{
getCharacterizationResultSet()
}
\description{
Get Characterization Result Set
}
