% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortCharacterization.R
\name{getCharacterizationSection}
\alias{getCharacterizationSection}
\title{Get Characterization Results Section}
\usage{
getCharacterizationSection()
}
\description{
Get Characterization Results Section
}
