% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{getCohortDefinitionSection}
\alias{getCohortDefinitionSection}
\title{Get Cohort Definition Section}
\usage{
getCohortDefinitionSection()
}
\description{
Obtains a report section describing the cohorts selected: 
- The print friendly definitions
- The counts
}
